Networks reuse policy
